修改安装脚本
更新 lazyvim 配置

Update lazyvim

Update lazyvim

修改 neovide 配置

修改路径

Neovide 配置

Neovide 配置

Update lazyvim

Update proxy

Update lazyvim

修改代理端口

修正安装脚本的错误

Update lazyvim

Update lazyvim

Update 20240218

添加 providers 开关

修改 JAVA_HOME

Remove old vim config

Update

+++ install.sh
@@ -1,31 +1,53 @@
-#!/bin/bash
+#!/usr/bin/env bash
+# Copyright (c) 2024-present Verdana Mu <verdana.cn@gmail.com>
+# This source code is licensed under the MIT license found in the license file.
 
-## Prepare proxy first if is WSL
+# nounset: Treat unset variables and parameters as an error when performing parameter expansion
+# errexit: Exit immediately if any command exits with a non-zero status
+set -o nounset -o errexit
 
-#if grep -qEi "(Microsoft|WSL)" /proc/version &> /dev/null ; then
-#fi
+# Workdir
+cd "$(dirname "${BASH_SOURCE}")"
+CWD=$(pwd)
 
-## Home directory
-cd ~
+# Homebrew
+if which brew >/dev/null; then
+	read -p "Install Homebrew packages? (y/n) " -n 1
+	echo ""
+	if [[ $REPLY =~ ^[Yy]$ ]]; then
+		brew bundle --no-lock --file "$CWD/Brewfile"
+	fi
+else
+	printf "\nHomebrew not installed! Skipping package installation..\n\n"
+fi
 
-## Remove old symlinks
-rm -rf .vim
-rm -rf .vimrc
+exit 0
 
-cp dotfiles/app/git/config .gitconfig
+## Copy gitconfig
+cp ~/dotfiles/git/config .gitconfig
 
 ## Create .config if missing
 test -d .config || mkdir -p .config
 
-ln -sf ~/dotfiles/fish .config/fish
-ln -sf ~/dotfiles/app/nvim .config/nvim
+function install_symlinks() {
+	test -L .vim && rm -rf .vim
+	test -L .vimrc && rm -rf .vimrc
+	test -L .config/fish && rm -rf .config/fish
+	test -L .config/nvim && rm -rf .config/nvim
 
-## Install fish theme
-curl git.io/pure-fish --location --output /tmp/pure_installer.fish
-fish -c "source /tmp/pure_installer.fish; and install_pure"
-
-function clean_symlinks() {
-	rm -rf .vim
-	rm -rf .vimrc
-	# test if .vim is symlink, then remove it
+	ln -sf ~/dotfiles/fish ~/.config/fish
+	ln -sf ~/dotfiles/vim ~/.config/nvim
 }
+
+function install_pure() {
+	curl git.io/pure-fish --location --output /tmp/pure_installer.fish
+	if [ $? -eq 0 ]; then
+		fish -c "source /tmp/pure_installer.fish; and install_pure"
+	else
+		echo "Download failed, please check your network"
+		exit -1
+	fi
+}
+
+install_pure
+install_symlinks
